+++ Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("NiumaClaw.Team")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.0.14.0")]
-[assembly: AssemblyInformationalVersion("1.0.14+native-desktop-download-fix.20260701")]
+[assembly: AssemblyFileVersion("1.0.15.0")]
+[assembly: AssemblyInformationalVersion("1.0.15+native-desktop-app-fix.20260701")]
 [assembly: AssemblyProduct("NiumaClaw.Team")]
 [assembly: AssemblyTitle("NiumaClaw.Team")]
-[assembly: AssemblyVersion("1.0.14.0")]
+[assembly: AssemblyVersion("1.0.15.0")]